requirements.txt
boto3==1.24.31
pandas==1.4.3
psycopg2-binary==2.9.3
redshift_connector==2.0.908
sphinx==5.0.2
sphinx_rtd_theme==1.0.0
